I wrapped my calculations in float

+++ python_introduction/finance_calculator.py
@@ -3,12 +3,12 @@ monthly_income = input("Enter your monthly income: ")
 monthly_expenses = input("Enter your montly expenses: ")
 
 # let us calculate the monthly saving while converting the results from the variables to integers.
-monthly_savings = int(monthly_income) - int(monthly_expenses)
+monthly_savings = float(int(monthly_income) - int(monthly_expenses))
 
 print("your current monthly savings:", "$", monthly_savings)
 
 # let us calculate the projected annual savings for one year
 
-projected_savings = monthly_savings * 12 + (monthly_savings * 12 * 0.05)
+projected_savings = float(monthly_savings * 12 + (monthly_savings * 12 * 0.05))
 
 print ("projected savings after one year, with interest, is: ", "$",projected_savings)
